refactor(mobile): split CloudTasks.tsx 743->212 LOC (Cluster A2)
CLAUDE.md mandates components <400 LOC. CloudTasks.tsx was 743. Behavior-
preserving extraction into co-located cloud-tasks/ modules:
- task-format.ts: STATUS_CONFIG, AGENT_COLORS + the pure formatters
  (formatRelativeTime/formatCost/computeProgress) — now unit-testable.
- rowToTask.ts: the snake_case -> camelCase row mapper.
- StatusBadge.tsx / ProgressBar.tsx / TaskCard.tsx: the sub-components.
- useCloudTasks.ts: the data + Supabase Realtime subscription + progress
  tick + optimistic cancel hook.
- CloudTasks.tsx (212): the presentational shell (loading/empty/list +
  detail modal) consuming the hook.

Tests: +11 for the extracted formatters + row mapper (edge cases: cap at 95%,
null session_id, hour pluralization). Existing cloud-tasks-screen.test.tsx
still passes; full mobile suite 1726 green = behavior preserved. typecheck +
lint clean; a11y guard passes (all extracted Pressables keep their labels).
